completed exercise 4.15-4.22

+++ part4/bloglist-backend/controllers/blog.js
@@ -1,64 +1,58 @@
-const blogRouter = require('express').Router()
+const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
-blogRouter.get('/', (request, response) => {
-  Blog.find({}).then(blogs => {
-    response.json(blogs)
-  })
-})
-
-blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-  if (!body.title || !body.url) {
-      return response.status(400).json({ error: 'title or url missing' })
-    }
+  const user = request.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
+    user: user._id
   })
-  try {
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
-  } catch (error) {
-    next(error)
-  }
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
-blogRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
-  } catch (error) {
-    next(error)
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
   }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'only the creator can delete this blog' })
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
+  response.status(204).end()
 })
 
-blogRouter.put('/:id', async (request, response, next) => {
-  const body = request.body
+blogsRouter.put('/:id', middleware.userExtractor, async (request, response) => {  const body = request.body
+
   const blog = {
-      likes: body.likes,
-      title: body.title,
-      author: body.author,
-      url: body.url
-    }
-  try {
-    const updatedBlog =
-      await Blog.findByIdAndUpdate(
-        request.params.id, 
-        blog, 
-        { returnDocument: 'after' }
-      )
-    response.json(updatedBlog)
-  } catch (error) {
-    next(error)
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes
   }
+
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+  response.json(updatedBlog)
 })
 
-module.exports = blogRouter
+module.exports = blogsRouter
